update validate fixbug avatar event_edit

+++ app/view/event_edit/EventEditInput.php
@@ -13,44 +13,44 @@
     <form method="post" <?php echo 'action="../eventEditInput/' . $event_id . '"' ?> enctype="multipart/form-data">
         <button id="back-home" name="back-home" >Quay lại</button><br>
 
-        <?php if (isset($validate) && $validate['name'] != '') {
-            if (array_key_exists('name', $validate)) {
+        <?php if (isset($validate) ) {
+            if (array_key_exists('name', $validate) && $validate['name'] != '') {
                 echo "<span class='error'>" . $validate['name'] . "</span>";
             }
-            if(array_key_exists('name_length', $validate)){
+            if(array_key_exists('name_length', $validate) && $validate['name_length'] != ''){
                 echo '<span class="error">' . $validate['name_length'] .'</span>';
             }
         } ?>
         <label class="input-form" for="name">Tên sự kiện</label>
         <input id="name" type="text" name="event_name" maxlength="" value="<?php echo $event_name; ?>"><br>
 
-        <?php if (isset($validate) && $validate['slogan'] != '' ) {
-            if (array_key_exists('slogan', $validate)) {
+        <?php if (isset($validate) ) {
+            if (array_key_exists('slogan', $validate) && $validate['slogan'] != '' ) {
                 echo "<span class='error'>" . $validate['slogan'] . "</span>";
             }
-            if(array_key_exists('slogan_length', $validate)){
+            if(array_key_exists('slogan_length', $validate) && $validate['slogan_length'] != '' ){
                 echo '<span class="error">' . $validate['slogan_length'] .'</span>';
             }
         } ?>
         <label class="input-form" for="slogan">Slogan</label>
         <input id="slogan" type="text" name="event_slogan" maxlength="" value="<?php echo $event_slogan; ?>"><br>
 
-        <?php if (isset($validate)  && $validate['leader'] != ''  ) {
-            if (array_key_exists('leader', $validate)) {
+        <?php if (isset($validate)   ) {
+            if (array_key_exists('leader', $validate) && $validate['leader'] != '' ) {
                 echo "<span class='error'>" . $validate['leader'] . "</span>";
             }
-            if(array_key_exists('leader_length', $validate)){
+            if(array_key_exists('leader_length', $validate) && $validate['leader_length'] != '' ){
                 echo '<span class="error">' . $validate['leader_length'] .'</span>';
             }
         } ?>
         <label class="input-form" for="leader">Leader</label>
         <input id="leader" type="text" name="event_leader" maxlength="" value="<?php echo $event_leader; ?>"><br>
 
-        <?php if (isset($validate)  && $validate['description'] != ''  ) {
-            if (array_key_exists('description', $validate)) {
+        <?php if (isset($validate)   ) {
+            if (array_key_exists('description', $validate) && $validate['description'] != '' ) {
                 echo "<span class='error'>" . $validate['description'] . "</span>";
             }
-            if(array_key_exists('description_length', $validate)){
+            if(array_key_exists('description_length', $validate) && $validate['description_length'] != '' ){
                 echo '<span class="error">' . $validate['description_length'] .'</span>';
             }
         } ?>
